+++ src/app/common/common.service.js
@@ -53,8 +53,8 @@
             getEditions: getEditions,
             getEducation: getEducation,
             getMeaningfulUseUsersAccurateAsOfDate: getMeaningfulUseUsersAccurateAsOfDate,
-            getNotificationRecipients: getNotificationRecipients,
-            getNotificationReportTypes: getNotificationReportTypes,
+            getSubscriptionRecipients: getSubscriptionRecipients,
+            getSubscriptionReportTypes: getSubscriptionReportTypes,
             getPractices: getPractices,
             getProduct: getProduct,
             getProductActivity: getProductActivity,
@@ -695,11 +695,11 @@
         /*
          * Email notification services
          */
-        function getNotificationReportTypes () {
+        function getSubscriptionReportTypes () {
             return simpleApiCall('/data/notification_types');
         }
 
-        function getNotificationRecipients () {
+        function getSubscriptionRecipients () {
             return simpleApiCall('/notifications/recipients');
         }