+++ RELEASE_NOTES.md
@@ -10,7 +10,19 @@ _Date TBD_
 
 ### New features
 * Added 'download definition file' download option
-* Added "API Criteria" collection
+* Triggers retired items on main filter when retired items are selected
+  * Trigger: Selecting any of:
+    * "Retired" status
+    * "2011" edition
+    * Retired ACB
+    * 2011 Criteria
+    * 2014 CQM
+  * Turns on all of:
+    * "Retired" status
+    * "2011" edition
+    * All retired ACBs
+* Changed admin Notifications section to just "Subscription Management"
+* Added "API Documentation" collection
 
 ---